requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
fonttools==4.49.0
greenlet==3.0.3
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
psycopg2==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
